Move some methods from Template class to HTMLEmail
Things like making links out of URLs and escaping ‘GOV.UK’ feel more
closely associated with a HTML email than a Notify template. The code
for the `Template` class was getting pretty massive too.

Breaking changes because it removes some methods on the `Template` class
(which probably aren’t being publicly used, but still).
