+++ notifications_utils/version.py
@@ -1,1 +1,1 @@
-__version__ = '6.4.0'
+__version__ = '7.0.0'
